backend/requirements.txt
# backend/requirements.txt

fastapi
uvicorn
pymongo
python-bson
passlib[bcrypt]
bcrypt==3.2.0
python-jose[cryptography]
pydantic
pydantic-settings
celery[redis]
redis
pinecone-client>=3.0.0,<4.0.0
python-dotenv
google-generativeai
cohere
anthropic
eventlet  # <-- ADD THIS LINE